requirements.txt
jinja2
gitpython
requests
lxml
bs4
toml
alive-progress
buildbot
